+++ src/sparseml/transformers/__init__.py
@@ -12,105 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""
-Tools for integrating SparseML with transformers training flows
-"""
-
-# flake8: noqa
-
-import logging as _logging
-
-
-try:
-    import transformers as _transformers
-
-    # triggers error if neuralmagic/transformers is not installed
-    assert _transformers.NM_INTEGRATED
-    _transformers_import_error = None
-except Exception as _transformers_import_err:
-    _transformers_import_error = _transformers_import_err
-
-
-_LOGGER = _logging.getLogger(__name__)
-_NM_TRANSFORMERS_TAR_TEMPLATE = (
-    "https://github.com/neuralmagic/transformers/releases/download/"
-    "{version}/transformers-4.18.0.dev0-py3-none-any.whl"
-)
-_NM_TRANSFORMERS_NIGHTLY = _NM_TRANSFORMERS_TAR_TEMPLATE.format(version="nightly")
-
-
-def _install_transformers_and_deps():
-
-    import subprocess as _subprocess
-    import sys as _sys
-
-    import sparseml as _sparseml
-
-    nm_transformers_release = (
-        "nightly" if not _sparseml.is_release else f"v{_sparseml.version_major_minor}"
-    )
-    transformers_requirement = _NM_TRANSFORMERS_TAR_TEMPLATE.format(
-        version=nm_transformers_release
-    )
-    try:
-        _subprocess.check_call(
-            [
-                _sys.executable,
-                "-m",
-                "pip",
-                "install",
-                transformers_requirement,
-                "datasets<=1.18.4",
-                "sklearn",
-                "seqeval",
-            ]
-        )
-
-        import transformers as _transformers
-
-        _LOGGER.info("sparseml-transformers and dependencies successfully installed")
-    except Exception:
-        raise ValueError(
-            "Unable to install and import sparseml-transformers dependencies check "
-            "that transformers is installed, if not, install via "
-            f"`pip install {_NM_TRANSFORMERS_NIGHTLY}`"
-        )
-
-
-def _check_transformers_install():
-    if _transformers_import_error is not None:
-        import os
-
-        if os.getenv("NM_NO_AUTOINSTALL_TRANSFORMERS", False):
-            _LOGGER.warning(
-                "Unable to import, skipping auto installation "
-                "due to NM_NO_AUTOINSTALL_TRANSFORMERS"
-            )
-            # skip any further checks
-            return
-        else:
-            _LOGGER.warning(
-                "sparseml-transformers installation not detected. Installing "
-                "sparseml-transformers dependencies if transformers is already "
-                "installed in the environment, it will be overwritten. Set "
-                "environment variable NM_NO_AUTOINSTALL_TRANSFORMERS to disable"
-            )
-            _install_transformers_and_deps()
-
-    # re check import after potential install
-    try:
-        import transformers as _transformers
-
-        assert _transformers.NM_INTEGRATED
-    except Exception:
-        _LOGGER.warning(
-            "the neuralmagic fork of transformers may not be installed. it can be "
-            "installed via "
-            f"`pip install {_NM_TRANSFORMERS_NIGHTLY}`"
-        )
-
-
-_check_transformers_install()
-
 from .export import *
 from .hot_patching import *
